requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
psycopg2-binary
SQLAlchemy
alembic
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
httpx
python-multipart
